New asyncio backend
Dear diary,
Haven't committed in a while. I'm not sorry.

So now we are using asycio to create a concurrent backend
that spawns worker tasks to handle IO over concurrent sessions.
We are now tricking the server into thinking we are the Orichil
client so that we receive webclient map data. SUCCESS.

The asyncio backend and the gui will run on separate threads and
communicate using q_manager, which is just a central manager for
creating and accessing queues.

I have decided to use the logging module to smash myself with info
allowing better, faster, and live debugging info. HIGH FIVE.